backend/requirements.txt
djangorestframework
djangorestframework-simplejwt
python-dateutil